requirements.txt
pandas==1.5.3
requests==2.31.0
python-dotenv==1.0.0
clickhouse-driver==0.2.10